feat: make python client an install package
This will allow to install it directly in jupyter docker image:

https://github.com/jolibrain/dd_platform_docker/pull/39
