+++ .gitignore
@@ -5,5 +5,4 @@ build/*
 !build/get_libs.sh
 model/
 models/
-clients/
 tools/build-cpp-netlib/
