+++ clients/python/README.md
@@ -10,10 +10,17 @@ or system-wide, e.g. on Ubuntu:
 sudo apt-get install python-requests
 ```
 
-Then either copy it into your Python code repository or point your `PYTHONPATH` to it:
+Then install deepdetect python client from deepdetect source code:
 
 ```
-export PYTHONPATH=/path/to/deepdetect/clients/python/:$PYTHONPATH
+cd deepdetect_src/clients/python
+pip install .
+```
+
+or directly with:
+
+```
+pip install git+https://github.com/deepdetect.git#egg=dd_client&subdirectory=dd_client
 ```
 
 The DD client may post images through they base64 representation if those are
@@ -21,4 +28,3 @@ not accessible to the deepdetect server. The base64 conversion is supported for
 images on the client filesystem or online (http/https), shall the url be only
 accessible to the client. A `base64` flag (default to `False`) is available in
 the `post_predict` method.
-
