+++ clients/python/setup.cfg
@@ -1,0 +1,25 @@
+[metadata]
+name = deepdetect
+home-page = https://github.com/jolibrain/deepdetect
+summary = DeepDetect API client
+author = Jolibrain
+author-email = emmanuel.benazera@jolibrain.com
+classifier =
+    Intended Audience :: Information Technology
+    License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)
+    Operating System :: POSIX
+    Programming Language :: Python
+    Programming Language :: Python :: 2
+    Programming Language :: Python :: 2.7
+
+[options]
+packages =
+    dd_client
+
+install_requires =
+    requests
+
+[options.entry_points]
+console_scripts =
+    dd_client = dd_client:main
+    dd_bench = dd_client.bench:main
