+++ clients/python/setup.py
@@ -1,0 +1,6 @@
+
+import setuptools
+
+setuptools.setup(
+    setup_requires=["setuptools>=30.3.0"],
+)
